# Use custom reflection for decoding structs