--- a/gocql_backend.go
+++ b/gocql_backend.go
@@ -8,23 +8,25 @@ type goCQLBackend struct {
 	session *gocql.Session
 }
 
-func (cb goCQLBackend) Query(stmt Statement) ([]map[string]interface{}, error) {
-	return cb.QueryWithOptions(Options{}, stmt)
+func (cb goCQLBackend) Query(stmt Statement, scanner Scanner) error {
+	return cb.QueryWithOptions(Options{}, stmt, scanner)
 }
 
-func (cb goCQLBackend) QueryWithOptions(opts Options, stmt Statement) ([]map[string]interface{}, error) {
+func (cb goCQLBackend) QueryWithOptions(opts Options, stmt Statement, scanner Scanner) error {
 	qu := cb.session.Query(stmt.Query(), stmt.Values()...)
 	if opts.Consistency != nil {
 		qu = qu.Consistency(*opts.Consistency)
 	}
+
 	iter := qu.Iter()
-	ret := []map[string]interface{}{}
-	m := &map[string]interface{}{}
-	for iter.MapScan(*m) {
-		ret = append(ret, *m)
-		m = &map[string]interface{}{}
+	_, err := scanner.ScanAll(iter)
+	errClose := iter.Close()
+
+	if err != nil {
+		return err
 	}
-	return ret, iter.Close()
+
+	return errClose
 }
 
 func (cb goCQLBackend) Execute(stmt Statement) error {