--- a/interfaces.go
+++ b/interfaces.go
@@ -277,10 +277,10 @@ type Scannable interface {
 
 // Scanner encapsulates a scanner which scans rows from a GoCQL iterator.
 type Scanner interface {
-	// ScanAll takes in a Scannale iterator found in GoCQL and scans until
+	// ScanIter takes in a Scannale iterator found in GoCQL and scans until
 	// the iterator giveth no more. It number of rows read and an optional
 	// error if anything goes wrong
-	ScanAll(iter Scannable) (int, error)
+	ScanIter(iter Scannable) (int, error)
 }
 
 // QueryExecutor actually executes the queries - this is mostly useful for testing/mocking purposes,