--- a/map_table.go
+++ b/map_table.go
@@ -5,13 +5,13 @@ type mapT struct {
 	idField string
 }
 
-func (m *mapT) Table() Table                     { return m.t }
-func (m *mapT) Create() error                    { return m.Table().Create() }
-func (m *mapT) CreateIfNotExist() error          { return m.Table().CreateIfNotExist() }
-func (m *mapT) Name() string                     { return m.Table().Name() }
-func (m *mapT) Recreate() error                  { return m.Table().Recreate() }
-func (m *mapT) CreateStatement() (string, error) { return m.Table().CreateStatement() }
-func (m *mapT) CreateIfNotExistStatement() (string, error) {
+func (m *mapT) Table() Table                        { return m.t }
+func (m *mapT) Create() error                       { return m.Table().Create() }
+func (m *mapT) CreateIfNotExist() error             { return m.Table().CreateIfNotExist() }
+func (m *mapT) Name() string                        { return m.Table().Name() }
+func (m *mapT) Recreate() error                     { return m.Table().Recreate() }
+func (m *mapT) CreateStatement() (Statement, error) { return m.Table().CreateStatement() }
+func (m *mapT) CreateIfNotExistStatement() (Statement, error) {
 	return m.Table().CreateIfNotExistStatement()
 }