This PR replaces the use of the [`mapstructure`](https://github.com/mitchellh/mapstructure) library with custom reflection which is much more optimised for `gocassa` internals. Doing so leads to much more performant decoding characteristics. 

**⚠️ Given the extensive changes in the interface, this PR bumps the version of gocassa to v2.0.0**

### Motivation

The motivation for this PR came from profiling our services. Turns out quite a lot of services which use Cassandra at scale were spending 50-60%+ of their CPU within gocassa's `decodeResult` step (with a sizeable ~10% additionally spent in `MapScan`). This is a direct invocation of `mapstructure`.

Given I was going to tweak the interfaces, I took the liberty of cleaning up and adding a `Statement` object instead of ferrying around a query string and list of values. This can also encapsulate field name information under the hood which is used in our scanner.

### Change

This patch essentially hooks into the `gocql` Scanner directly rather than going through a map for `SELECT` queries. 

Previously we had two levels of reflection when reading via gocassa:
- A conversion from the Cassandra type into a `map[string]interface{}` within `gocql`. This used the `MapScan` function in `gocql` and thus results in a map per result allocated and populated
- A conversion using `decodeResult` in `gocassa` to convert each map into the target struct

Now we have a direct reflection into struct by leveraging the ordering of fields which gocassa also knows about since gocassa is what constructs the `SELECT` queries in the first place! 

### Benchmarks

**`gocassa` on `master`** (the benchmark test is at [`96f213d`](https://github.com/monzo/gocassa/commit/96f213d4c3a1993cd38011d32044771074c984ad)):
```
goos: darwin
goarch: amd64
pkg: github.com/monzo/gocassa
BenchmarkDecodeBlogSliceNoBody-4       	    3000	    480115 ns/op	   73486 B/op	    1687 allocs/op
BenchmarkDecodeBlogStruct-4            	     100	  12456151 ns/op	  375150 B/op	   30385 allocs/op
BenchmarkDecodeBlogStructEmptyBody-4   	  100000	     22711 ns/op	    3489 B/op	      68 allocs/op
BenchmarkDecodeAlphaSlice-4            	    1000	   1472692 ns/op	  306253 B/op	    5007 allocs/op
BenchmarkDecodeAlphaStruct-4           	   20000	     68542 ns/op	   14407 B/op	     174 allocs/op
PASS
ok  	github.com/monzo/gocassa	9.605s
```

These values are very high and it doesn't even account for the map allocation/population! I did some profiling and it turns out the [`DecodeHook` functionality we use](3fa24fca0346248338401cf2f93bdf938d55ae19) to convert from the `gocql` type of various integer types and also to support `UnmarshalCQL` functionality adds a very large overhead! I got rid of them briefly and ran another test.

**`gocassa` on `master` without DecodeHook functions** (the benchmark test is at [`96f213d`](https://github.com/monzo/gocassa/commit/96f213d4c3a1993cd38011d32044771074c984ad)):

```
goos: darwin
goarch: amd64
pkg: github.com/monzo/gocassa
BenchmarkDecodeBlogSliceNoBody-4       	   10000	    181272 ns/op	   73410 B/op	    1685 allocs/op
BenchmarkDecodeBlogStruct-4            	    1000	   1836223 ns/op	  374693 B/op	   30383 allocs/op
BenchmarkDecodeBlogStructEmptyBody-4   	  200000	      7485 ns/op	    3425 B/op	      66 allocs/op
BenchmarkDecodeAlphaSlice-4            	    2000	    720789 ns/op	  306197 B/op	    5005 allocs/op
BenchmarkDecodeAlphaStruct-4           	   50000	     35752 ns/op	   14339 B/op	     172 allocs/op
PASS
ok  	github.com/monzo/gocassa	10.730s
```

Still quite high but there has been noticeable improvement, especially in `BenchmarkDecodeBlogStruct` where decoding a single struct with a `[]byte` populated was making the per operation latency 9x slower. Note that we can't just get rid of these functions in real use because otherwise our integer types will be wrong. You don't even need to be using any of this functionality, just the conversion to execute these decode hooks adds a high amount of latency.

**`gocassa` with this PR on decoding into structs using the iterator directly:**
```
goos: darwin
goarch: amd64
pkg: github.com/monzo/gocassa
BenchmarkDecodeBlogSliceNoBody-4       	   50000	     34113 ns/op	   21280 B/op	     235 allocs/op
BenchmarkDecodeBlogStruct-4            	  500000	      3403 ns/op	    2784 B/op	      32 allocs/op
BenchmarkDecodeBlogStructEmptyBody-4   	  500000	      3281 ns/op	    2784 B/op	      32 allocs/op
BenchmarkDecodeAlphaSlice-4            	   10000	    103263 ns/op	   43836 B/op	     810 allocs/op
BenchmarkDecodeAlphaStruct-4           	  200000	      9768 ns/op	    9019 B/op	      92 allocs/op
PASS
ok  	github.com/monzo/gocassa	9.222s
```

This is a 5-10x+ improvement in latency per operation and significantly reduced memory allocations by hooking into gocql directly. If you are using byte slices or other similar data structures within your target unmarshal struct, the improvement will be massive (upwards of hundreds of times faster). 

```
benchmark                                old ns/op     new ns/op     delta
BenchmarkDecodeBlogSliceNoBody-4         480115        34113         -92.89%
BenchmarkDecodeBlogStruct-4              12456151      3403          -99.97%
BenchmarkDecodeBlogStructEmptyBody-4     22711         3281          -85.55%
BenchmarkDecodeAlphaSlice-4              1472692       103263        -92.99%
BenchmarkDecodeAlphaStruct-4             68542         9768          -85.75%

benchmark                                old allocs     new allocs     delta
BenchmarkDecodeBlogSliceNoBody-4         1687           235            -86.07%
BenchmarkDecodeBlogStruct-4              30385          32             -99.89%
BenchmarkDecodeBlogStructEmptyBody-4     68             32             -52.94%
BenchmarkDecodeAlphaSlice-4              5007           810            -83.82%
BenchmarkDecodeAlphaStruct-4             174            92             -47.13%

benchmark                                old bytes     new bytes     delta
BenchmarkDecodeBlogSliceNoBody-4         73486         21280         -71.04%
BenchmarkDecodeBlogStruct-4              375150        2784          -99.26%
BenchmarkDecodeBlogStructEmptyBody-4     3489          2784          -20.21%
BenchmarkDecodeAlphaSlice-4              306253        43836         -85.69%
BenchmarkDecodeAlphaStruct-4             14407         9019          -37.40%
```

This PR isn't adding any additional hidden work on the `gocql` side because this iteration cost is already being spent by using `MapScan` which does the same work (in fact, as of writing this, it uses the `gocql.Iter.Scan()` under the hood).