@suhailpatel 2019-06-24
@milesbxf I did consider very heavily code generation but that got very complicated. This doesn't stop it though so potentially as another update if needed it'd be great!